folder content downloading (F10)
git-svn-id: https://svn.code.sf.net/p/gopherus/code/trunk@133 767d67df-c2ba-418b-950d-eab06a56d38d

+++ history.h
@@ -15,7 +15,7 @@ struct historytype {
   unsigned short port;
   char protocol;
   char itemtype;
-  int displaymemory[2];  /* used by some display plugins to remember how the item was displayed. this is always initialized to -1 values */
+  long displaymemory[2];  /* used by some display plugins to remember how the item was displayed. this is always initialized to -1 values */
 };
 
 /* remove the last visited page from history (goes back to the previous one) */
